+++ .gitignore
@@ -2,7 +2,7 @@
 ## files generated by popular Visual Studio add-ons.
 ##
 ## Get latest from https://github.com/github/gitignore/blob/master/VisualStudio.gitignore
-
+images/logo_original.png
 # User-specific files
 *.rsuser
 *.suo
