+++ Nidhogg.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32510.428
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.34024.191
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Nidhogg", "Nidhogg\Nidhogg.vcxproj", "{13C57810-FF18-4258-ABC9-935040A54F0B}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "NidhoggClient", "NidhoggClient\NidhoggClient.vcxproj", "{67F619F9-91AB-437B-92C7-3DE51D4110E0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -41,6 +43,22 @@ Global
 		{13C57810-FF18-4258-ABC9-935040A54F0B}.Release|x86.ActiveCfg = Release|Win32
 		{13C57810-FF18-4258-ABC9-935040A54F0B}.Release|x86.Build.0 = Release|Win32
 		{13C57810-FF18-4258-ABC9-935040A54F0B}.Release|x86.Deploy.0 = Release|Win32
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Debug|ARM.ActiveCfg = Debug|x64
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Debug|ARM.Build.0 = Debug|x64
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Debug|ARM64.ActiveCfg = Debug|x64
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Debug|ARM64.Build.0 = Debug|x64
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Debug|x64.ActiveCfg = Debug|x64
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Debug|x64.Build.0 = Debug|x64
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Debug|x86.ActiveCfg = Debug|Win32
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Debug|x86.Build.0 = Debug|Win32
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Release|ARM.ActiveCfg = Release|x64
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Release|ARM.Build.0 = Release|x64
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Release|ARM64.ActiveCfg = Release|x64
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Release|ARM64.Build.0 = Release|x64
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Release|x64.ActiveCfg = Release|x64
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Release|x64.Build.0 = Release|x64
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Release|x86.ActiveCfg = Release|Win32
+		{67F619F9-91AB-437B-92C7-3DE51D4110E0}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
