+++ NidhoggClient/pch.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <iostream>
+#include <Windows.h>
+#include <vector>
+#include <string>
+#include <sddl.h>
+#pragma comment(lib, "advapi32.lib")
